fixed deletion when Exception launched at fetching

+++ app/src/com/constellio/app/modules/es/connectors/smb/jobs/SmbDeleteJob.java
@@ -1,6 +1,7 @@
 package com.constellio.app.modules.es.connectors.smb.jobs;
 
 import com.constellio.app.modules.es.connectors.smb.jobmanagement.SmbConnectorJob;
+import com.constellio.app.modules.es.connectors.smb.service.SmbFileDTO;
 import com.constellio.app.modules.es.connectors.smb.service.SmbRecordService;
 import com.constellio.app.modules.es.connectors.spi.Connector;
 import com.constellio.app.modules.es.model.connectors.DocumentSmbConnectorUrlCalculator;
@@ -34,20 +35,23 @@ public class SmbDeleteJob extends SmbConnectorJob {
     public void execute(Connector connector) {
         String url = jobParams.getUrl();
         try {
-            if (jobParams.getSmbUtils().isFolder(url)) {
-                ConnectorSmbFolder folderToDelete = jobParams.getSmbRecordService().getFolderFromCache(url, jobParams.getConnectorInstance());
-                if(folderToDelete != null) {
-                    jobParams.getSmbRecordService().removeFromCache(folderToDelete);
-                    deleteByUrl(folderToDelete.getUrl());
-                    DeleteEventOptions options = new DeleteEventOptions();
-                    jobParams.getEventObserver().deleteEvents(options, folderToDelete);
-                }
-            } else {
-                ConnectorSmbDocument documentToDelete = jobParams.getSmbRecordService().getDocumentFromCache(url, jobParams.getConnectorInstance());
-                if(documentToDelete != null) {
-                    jobParams.getEventObserver().deleteEvents(documentToDelete);
-                    jobParams.getSmbRecordService().removeFromCache(documentToDelete);
-                    deleteByUrl(url);
+            SmbFileDTO smbFileDTO = jobParams.getSmbShareService().getSmbFileDTO(url, false);
+            if(smbFileDTO.getStatus() == SmbFileDTO.SmbFileDTOStatus.DELETE_DTO) {
+                if (jobParams.getSmbUtils().isFolder(url)) {
+                    ConnectorSmbFolder folderToDelete = jobParams.getSmbRecordService().getFolderFromCache(url, jobParams.getConnectorInstance());
+                    if(folderToDelete != null) {
+                        jobParams.getSmbRecordService().removeFromCache(folderToDelete);
+                        deleteByUrl(folderToDelete.getUrl());
+                        DeleteEventOptions options = new DeleteEventOptions();
+                        jobParams.getEventObserver().deleteEvents(options, folderToDelete);
+                    }
+                } else {
+                    ConnectorSmbDocument documentToDelete = jobParams.getSmbRecordService().getDocumentFromCache(url, jobParams.getConnectorInstance());
+                    if(documentToDelete != null) {
+                        jobParams.getEventObserver().deleteEvents(documentToDelete);
+                        jobParams.getSmbRecordService().removeFromCache(documentToDelete);
+                        deleteByUrl(url);
+                    }
                 }
             }
         } catch (Exception e) {
